+++ packages/snaps-controllers/src/snaps/index.ts
@@ -40,8 +40,6 @@ export type {
   SnapControllerRemoveSnapsAction,
   SnapControllerDisconnectOriginAction,
   SnapControllerRevokeDynamicSnapPermissionsAction,
-  SnapControllerIncrementActiveReferencesAction,
-  SnapControllerDecrementActiveReferencesAction,
   SnapControllerGetAllSnapsAction,
   SnapControllerGetRunnableSnapsAction,
   SnapControllerGetPermittedSnapsAction,